Remove old auth types and move memory-card-specific types to libs/auth
