+++ libs/types/src/auth/index.ts
@@ -1,7 +1,3 @@
 export * from './auth';
 export * as DippedSmartCardAuth from './dipped_smart_card_auth';
 export * as InsertedSmartCardAuth from './inserted_smart_card_auth';
-
-/** Being phased out */
-export * as DippedSmartcardAuth from './dipped_smartcard_auth';
-export * as InsertedSmartcardAuth from './inserted_smartcard_auth';
